[refactor] Scrap API refactor ver 2
