+++ src/main/java/com/sunny/backend/common/CommonErrorCode.java
@@ -17,7 +17,8 @@ public enum CommonErrorCode {
     INVALID_FUTURE_DATE(HttpStatus.FORBIDDEN, "미래 날짜는 사용할 수 없습니다."),
     TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "만료된 토큰입니다."),
     TOKEN_INVALID(HttpStatus.FORBIDDEN, "권한이 없습니다."),
-    ALREADY_SCRAP(HttpStatus.BAD_REQUEST, "이미 스크랩한 게시글입니다.");
+    ALREADY_SCRAP(HttpStatus.BAD_REQUEST, "이미 스크랩한 게시글입니다."),
+    NOT_FOUND_SCRAP(HttpStatus.NOT_FOUND, "스크랩 게시글을 찾을 수 없습니다.");
 
     private final HttpStatus httpStatus;
     private final String message;
