+++ src/main/java/com/sunny/backend/scrap/domain/Scrap.java
@@ -1,5 +1,6 @@
-package com.sunny.backend.entity;
+package com.sunny.backend.scrap.domain;
 
+import com.sunny.backend.entity.Community;
 import com.sunny.backend.user.Users;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
