+++ src/main/java/com/sunny/backend/scrap/repository/ScrapRepository.java
@@ -1,8 +1,7 @@
-package com.sunny.backend.repository;
+package com.sunny.backend.scrap.repository;
 
-import com.sunny.backend.dto.response.ScrapResponse;
 import com.sunny.backend.entity.Community;
-import com.sunny.backend.entity.Scrap;
+import com.sunny.backend.scrap.domain.Scrap;
 import com.sunny.backend.user.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
