+++ src/main/java/com/sunny/backend/service/MyPageService.java
@@ -7,8 +7,8 @@ import com.sunny.backend.dto.response.comment.CommentResponse;
 import com.sunny.backend.dto.response.community.CommunityResponse;
 import com.sunny.backend.entity.Comment;
 import com.sunny.backend.entity.Community;
-import com.sunny.backend.entity.Scrap;
-import com.sunny.backend.repository.ScrapRepository;
+import com.sunny.backend.scrap.domain.Scrap;
+import com.sunny.backend.scrap.repository.ScrapRepository;
 import com.sunny.backend.repository.comment.CommentRepository;
 import com.sunny.backend.repository.community.CommunityRepository;
 import com.sunny.backend.security.userinfo.CustomUserPrincipal;
