+++ src/main/java/com/sunny/backend/user/Users.java
@@ -2,6 +2,7 @@ package com.sunny.backend.user;
 
 import com.sunny.backend.consumption.domain.Consumption;
 import com.sunny.backend.save.domain.Save;
+import com.sunny.backend.scrap.domain.Scrap;
 import java.util.ArrayList;
 import java.util.List;
